Removing require-await rule
This rule is generally nice but also causes us some red flags when we want to
write Promise based functions that themselves don't do asynchronous work. This
mostly manifests in Lambda handlers, but can happen in other places where we
have a lot of Promise based functions and want to write them as async for
consistency.